+++ src/core_modules/capture-core/components/Pages/MainPage/EventWorkingListsInit/EventWorkingListsInit.container.type.js
@@ -1,0 +1,6 @@
+// @flow
+export type Props = $ReadOnly<{|
+    programId: string,
+    programStageId: string,
+    orgUnitId: string
+|}>;