+++ src/core_modules/capture-core/metaData/index.js
@@ -24,7 +24,6 @@ export { SearchGroup } from './SearchGroup';
 export { InputSearchGroup } from './InputSearchGroup';
 export { TrackedEntityType, TeiRegistration } from './TrackedEntityType';
 export { SystemSettings } from './SystemSettings';
-
 // helpers
 export {
     getScopeFromScopeId,
@@ -41,4 +40,5 @@ export {
     getScopeInfo,
     programTypes,
     scopeTypes,
+    getProgramThrowIfNotFound,
 } from './helpers';